+++ cmd/autoseed/testdata/schema/order.go
@@ -1,0 +1,19 @@
+package schema
+
+import (
+	"entgo.io/ent"
+	"entgo.io/ent/schema/edge"
+)
+
+// Order has a required belongs-to Customer, the inverse of Customer's
+// orders edge.
+type Order struct {
+	ent.Schema
+}
+
+// Edges returns Order's edges.
+func (Order) Edges() []ent.Edge {
+	return []ent.Edge{
+		edge.From("customer", Customer.Type).Ref("orders").Unique().Required(),
+	}
+}